Search and results made into component
